Homework 4/services/analysis/lstm/requirements.txt
setuptools
aiohttp
asyncpg
python-dotenv
tensorflow-cpu
numpy
scikit-learn
